Staging

+++ Modules/bookHandler.js
@@ -8,7 +8,11 @@ const bookHandler = {};
 
 // Function getBooks - is a method of bookHandler
 bookHandler.getBooks = function(request, expressResponse, next){
-  Book.find({})
+
+  // Set property of user email
+  let queryObject = {email: request.user.email};
+
+  Book.find(queryObject) // Find data by user email
     .then(data => {
       if (data) {
         expressResponse.status(200).send(data);
@@ -26,6 +30,7 @@ bookHandler.getBooks = function(request, expressResponse, next){
 
 // Function postBooks - method of bookHandler
 bookHandler.postBooks = function(request, expressResponse, next){
+
   const data = request.body; // Data sent from front-end to back-end gets received as req.body on the back-end
 
   if (!data || !data.title || !data.description) {
@@ -33,8 +38,11 @@ bookHandler.postBooks = function(request, expressResponse, next){
     return expressResponse.status(400).send('Invalid book data');
   }
 
-  // Create new book from data and send back confirmation if data created successfully
-  Book.create(data)
+  // Add user email property to book data
+  let bookDataWithEmailAdded = {...data, email: request.user.email};
+
+  // Create new book from data along with user email and send back confirmation if data created successfully
+  Book.create(bookDataWithEmailAdded)
     .then(createdBook => expressResponse.status(201).send(createdBook))
     .catch(err => {
       // Log error and send a 500 Internal Server Error response
@@ -47,12 +55,27 @@ bookHandler.postBooks = function(request, expressResponse, next){
 bookHandler.updateBooks = function(request, expressResponse, next){
   const {id} = request.params;
   const data = request.body;
+  const userEmail = request.user.email;
+
   // new - returns updated doc instead of old doc
   // overwrite - overwrites doc completely avoiding unwanted properties/side-effects
-  Book.findByIdAndUpdate(id, data, {new: true, overwrite: true})
-    .then(updatedBook => expressResponse.status(200).send(updatedBook))
+  Book.findById(id)
+    .then(book => {
+      if (!book) {
+        return expressResponse.status(404).send('Book not found');
+      }
+
+      if (book.email !== userEmail) {
+        return expressResponse.status(403).send('Unauthorized to update this book');
+      }
+
+      // Proceed with update if the user email matches
+      return Book.findByIdAndUpdate(id, data, { new: true, overwrite: true });
+    })
+    .then(updatedBook => {
+      expressResponse.status(200).send(updatedBook);
+    })
     .catch(err => {
-      // Log error and send a 500 Internal Server Error response
       console.error(err);
       expressResponse.status(500).send('Internal Server Error, updating book');
     });
@@ -61,10 +84,26 @@ bookHandler.updateBooks = function(request, expressResponse, next){
 // Funciton deleteBooks - method of bookHandler
 bookHandler.deleteBooks = function(request, expressResponse, next){
   const {id} = request.params;
-  Book.findByIdAndDelete(id)
-    .then(deletedBook => expressResponse.status(200).send(deletedBook)) // Send back confirmation of deleted data
+  const userEmail = request.user.email;
+
+  // First, find book and check if it belongs to the user by comparing email
+  Book.findById(id)
+    .then(book => {
+      if (!book) {
+        return expressResponse.status(404).send('Book not found');
+      }
+
+      if (book.email !== userEmail) {
+        return expressResponse.status(403).send('Unauthorized to delete this book');
+      }
+
+      // Proceed with deletion if the user email matches
+      return Book.findByIdAndDelete(id);
+    })
+    .then(deletedBook => {
+      expressResponse.status(200).send(deletedBook);
+    })
     .catch(err => {
-      // Log error and send a 500 Internal Server Error response
       console.error(err);
       expressResponse.status(500).send('Internal Server Error, deleting book');
     });
